+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   `StackReference` type.
 - Add CI system detection for Bitbucket Pipelines.
 - Add support for renaming resources via the `aliases` resource option.  Adding aliases allows new resources to match
-  resource from previous deployments which used different names, maintaining the identity of the resource and avoiding
+  resources from previous deployments which used different names, maintaining the identity of the resource and avoiding
   replacements or re-creation of the resource.
 
 ## 0.17.14 (Released May 28, 2019)
