+++ tests/integration/aliases/adopt_into_component/step1/index.ts
@@ -1,0 +1,18 @@
+// Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
+
+import * as pulumi from "@pulumi/pulumi";
+
+class Resource extends pulumi.ComponentResource {
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:Resource", name, {}, opts);
+    }
+}
+
+// Scenario #2 - adopt a resource into a component
+class Component extends pulumi.ComponentResource {
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:Component", name, {}, opts);
+    }
+}
+const res2 = new Resource("res2");
+const comp2 = new Component("comp2");
