+++ tests/integration/aliases/adopt_into_component/step1/package.json
