+++ tests/integration/aliases/adopt_into_component/step2/index.ts
@@ -1,0 +1,29 @@
+// Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
+
+import * as pulumi from "@pulumi/pulumi";
+
+class Resource extends pulumi.ComponentResource {
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:Resource", name, {}, opts);
+    }
+}
+
+// Scenario #2 - adopt a resource into a component.  The component author is the same as the component user, and changes
+// the component to be able to adopt the resource that was previously defined separately...
+class Component extends pulumi.ComponentResource {
+    resource: Resource;
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:Component", name, {}, opts);
+        // The resource creation was moved from top level to inside the component.
+        this.resource = new Resource(`${name}-child`, {
+            // With a new parent
+            parent: this,
+            // But with an alias provided based on knowing where the resource existing before - in this case at top
+            // level.  We use an absolute URN instead of a relative `Alias` because we are referencing a fixed resource
+            // that was in some arbitrary other location in the hierarchy prior to being adopted into this component.
+            aliases: [pulumi.createUrn("res2", "my:module:Resource")],
+        });
+    }
+}
+// The creation of the component is unchanged.
+const comp2 = new Component("comp2");
