+++ tests/integration/aliases/aliases_test.go
@@ -3,6 +3,7 @@
 package ints
 
 import (
+	"path"
 	"testing"
 
 	"github.com/pulumi/pulumi/pkg/testing/integration"
@@ -11,16 +12,22 @@ import (
 // TestAliases tests a case where a resource's name changes but it provides an `alias` pointing to the old URN to ensure
 // the resource is preserved across the update.
 func TestAliases(t *testing.T) {
-	integration.ProgramTest(t, &integration.ProgramTestOptions{
-		Dir:          "step1",
-		Dependencies: []string{"@pulumi/pulumi"},
-		Quick:        true,
-		EditDirs: []integration.EditDir{
-			{
-				Dir:             "step2",
-				Additive:        true,
-				ExpectNoChanges: true,
-			},
-		},
-	})
+	dirs := []string{"rename", "adopt_into_component", "rename_component", "rename_component_and_child", "retype_component"}
+	for _, dir := range dirs {
+		d := dir
+		t.Run(d, func(t *testing.T) {
+			integration.ProgramTest(t, &integration.ProgramTestOptions{
+				Dir:          path.Join(d, "step1"),
+				Dependencies: []string{"@pulumi/pulumi"},
+				Quick:        true,
+				EditDirs: []integration.EditDir{
+					{
+						Dir:             path.Join(d, "step2"),
+						Additive:        true,
+						ExpectNoChanges: true,
+					},
+				},
+			})
+		})
+	}
 }
