+++ tests/integration/aliases/rename/step1/Pulumi.yaml
@@ -1,3 +1,3 @@
-name: aliases
+name: aliases_rename
 description: A program that replaces a resource with a new name and alias.
 runtime: nodejs
