+++ tests/integration/aliases/rename_component/step1/stack2.json
@@ -1,0 +1,53 @@
+{
+    "version": 3,
+    "deployment": {
+        "manifest": {
+            "time": "2019-05-30T21:53:25.006537-07:00",
+            "magic": "e3fda2507d27c54305e184a60e1e88464cb1a3261d30512895fcb0e72ef8b3e0",
+            "version": "v0.17.14-dev.1559253032+g042f9045.dirty"
+        },
+        "secrets_providers": {
+            "type": "service",
+            "state": {
+                "url": "https://api.pulumi.com",
+                "owner": "lukehoban",
+                "project": "aliases_rename_component",
+                "stack": "asdsa"
+            }
+        },
+        "resources": [
+            {
+                "urn": "urn:pulumi:asdsa::aliases_rename_component::pulumi:pulumi:Stack::aliases_rename_component-asdsa",
+                "custom": false,
+                "type": "pulumi:pulumi:Stack"
+            },
+            {
+                "urn": "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree::newcomp3",
+                "custom": false,
+                "type": "my:module:ComponentThree",
+                "parent": "urn:pulumi:asdsa::aliases_rename_component::pulumi:pulumi:Stack::aliases_rename_component-asdsa",
+                "aliases": [
+                    "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree::comp3"
+                ]
+            },
+            {
+                "urn": "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree$my:module:Resource::otherchild",
+                "custom": false,
+                "type": "my:module:Resource",
+                "parent": "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree::newcomp3",
+                "aliases": [
+                    "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree$my:module:Resource::otherchild"
+                ]
+            },
+            {
+                "urn": "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree$my:module:Resource::newcomp3-child",
+                "custom": false,
+                "type": "my:module:Resource",
+                "parent": "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree::newcomp3",
+                "aliases": [
+                    "urn:pulumi:asdsa::aliases_rename_component::my:module:ComponentThree$my:module:Resource::comp3-child"
+                ]
+            }
+        ]
+    }
+}
