+++ tests/integration/aliases/rename_component_and_child/step1/Pulumi.yaml
@@ -1,0 +1,3 @@
+name: aliases_rename_component_and_child
+description: A program that replaces a resource with a new name and alias.
+runtime: nodejs
