+++ tests/integration/aliases/rename_component_and_child/step1/index.ts
@@ -1,0 +1,19 @@
+// Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
+
+import * as pulumi from "@pulumi/pulumi";
+
+class Resource extends pulumi.ComponentResource {
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:Resource", name, {}, opts);
+    }
+}
+
+// Scenario #5 - composing #1 and #3 and making both changes at the same time
+class ComponentFive extends pulumi.ComponentResource {
+    resource: Resource;
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:ComponentFive", name, {}, opts);
+        this.resource = new Resource("otherchild", {parent: this});
+    }
+}
+const comp5 = new ComponentFive("comp5");
