+++ tests/integration/aliases/retype_component/step1/index.ts
@@ -1,0 +1,19 @@
+// Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
+
+import * as pulumi from "@pulumi/pulumi";
+
+class Resource extends pulumi.ComponentResource {
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:Resource", name, {}, opts);
+    }
+}
+
+// Scenario #4 - change the type of a component
+class ComponentFour extends pulumi.ComponentResource {
+    resource: Resource;
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:ComponentFour", name, {}, opts);
+        this.resource = new Resource("otherchild", {parent: this});
+    }
+}
+const comp4 = new ComponentFour("comp4");
