+++ tests/integration/aliases/retype_component/step2/index.ts
@@ -1,0 +1,27 @@
+// Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
+
+import * as pulumi from "@pulumi/pulumi";
+
+const stackName = pulumi.getStack();
+const projectName = pulumi.getProject();
+
+class Resource extends pulumi.ComponentResource {
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        super("my:module:Resource", name, {}, opts);
+    }
+}
+
+// Scenario #4 - change the type of a component
+class ComponentFour extends pulumi.ComponentResource {
+    resource: Resource;
+    constructor(name: string, opts?: pulumi.ComponentResourceOptions) {
+        // Add an alias that references the old type of this resource...
+        const aliases = [{ type: "my:module:ComponentFour" }, ...((opts && opts.aliases) || [])];
+        // ..and then make the super call with the new type of this resource and the added alias.
+        super("my:differentmodule:ComponentFourWithADifferentTypeName", name, {}, { ...opts, aliases });
+        // The child resource will also pick up an implicit alias due to the new type of the component it is parented
+        // to.
+        this.resource = new Resource("otherchild", { parent: this });
+    }
+}
+const comp4 = new ComponentFour("comp4");
